allow users update profile picture

+++ lib/src/modules/profile/data/profile_source.dart
@@ -8,14 +8,27 @@ class ProfileDataSource {
 
   ProfileDataSource(this._authService, this._storageService);
 
-  Future<String> uploadProfilePicture(String path) async {
+  Future<String> updateProfilePicture(String path) async {
+    final downloadUrl = await _uploadProfilePicture(path);
+
+    await _authService.updateUserInfo(
+      data: {
+        ...?_authService.currentState?.userMetadata,
+        'profile_photo': downloadUrl,
+      },
+    );
+    return downloadUrl;
+  }
+
+  Future<String> _uploadProfilePicture(String path) async {
     final userImage = _authService.currentState!.userMetadata?['profile_photo'];
 
-    await _storageService.delete(
-      [userImage],
-      bucket: StorageBuckets.profilePhoto,
-    );
-
+    if (userImage != null) {
+      await _storageService.delete(
+        [userImage],
+        bucket: StorageBuckets.profilePhoto,
+      );
+    }
     final downloadUrl = await _storageService.uploadFile(
       path,
       bucket: StorageBuckets.profilePhoto,
